roles/modeling/fast/requirements.txt
pip==20.2.2
tensorflow==2.3.0
google-cloud-storage==1.30
pandas